Fix CI
Move common code to utils.js
Update node version in Travis config to work with  Travis CI.
