+++ src/regression.js
@@ -1,41 +1,6 @@
+const { round, deriveDataProperties } = require('./utils');
+
 const DEFAULT_OPTIONS = { order: 2, precision: 2, period: null };
-
-/**
-* Determine the coefficient of determination (r^2) of a fit from the observations
-* and predictions.
-*
-* @param {Array<Array<number>>} data - Pairs of observed x-y values
-* @param {Array<Array<number>>} results - Pairs of observed predicted x-y values
-*
-* @return {number} - The r^2 value, or NaN if one cannot be calculated.
-*/
-function determinationCoefficient(data, results) {
-  const predictions = [];
-  const observations = [];
-
-  data.forEach((d, i) => {
-    if (d[1] !== null) {
-      observations.push(d);
-      predictions.push(results[i]);
-    }
-  });
-
-  const sum = observations.reduce((a, observation) => a + observation[1], 0);
-  const mean = sum / observations.length;
-
-  const ssyy = observations.reduce((a, observation) => {
-    const difference = observation[1] - mean;
-    return a + (difference * difference);
-  }, 0);
-
-  const sse = observations.reduce((accum, observation, index) => {
-    const prediction = predictions[index];
-    const residual = observation[1] - prediction[1];
-    return accum + (residual * residual);
-  }, 0);
-
-  return 1 - (sse / ssyy);
-}
 
 /**
 * Determine the solution of a system of linear equations A * x = b using
@@ -85,21 +50,6 @@ function gaussianElimination(input, order) {
 }
 
 /**
-* Round a number to a precision, specificed in number of decimal places
-*
-* @param {number} number - The number to round
-* @param {number} precision - The number of decimal places to round to:
-*                             > 0 means decimals, < 0 means powers of 10
-*
-*
-* @return {numbr} - The number, rounded
-*/
-function round(number, precision) {
-  const factor = 10 ** precision;
-  return Math.round(number * factor) / factor;
-}
-
-/**
 * The set of all fitting methods
 *
 * @namespace
@@ -130,16 +80,11 @@ const methods = {
       round((gradient * x) + intercept, options.precision)]
     );
 
-    const points = data.map((point) => predict(point[0]));
-    const residuals = points.map((point, index) => [point[0], point[1] - data[index][1]]);
-
     return {
-      points,
-      residuals,
       predict,
       equation: [gradient, intercept],
-      r2: round(determinationCoefficient(data, points), options.precision),
       string: intercept === 0 ? `y = ${gradient}x` : `y = ${gradient}x + ${intercept}`,
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -167,16 +112,11 @@ const methods = {
       round(coeffA * Math.exp(coeffB * x), options.precision),
     ]);
 
-    const points = data.map((point) => predict(point[0]));
-    const residuals = points.map((point, index) => [point[0], point[1] - data[index][1]]);
-
     return {
-      points,
-      residuals,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA}e^(${coeffB}x)`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -202,16 +142,11 @@ const methods = {
       round(round(coeffA + (coeffB * Math.log(x)), options.precision), options.precision),
     ]);
 
-    const points = data.map((point) => predict(point[0]));
-    const residuals = points.map((point, index) => [point[0], point[1] - data[index][1]]);
-
     return {
-      points,
-      residuals,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA} + ${coeffB} ln(x)`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -238,16 +173,11 @@ const methods = {
       round(round(coeffA * (x ** coeffB), options.precision), options.precision),
     ]);
 
-    const points = data.map((point) => predict(point[0]));
-    const residuals = points.map((point, index) => [point[0], point[1] - data[index][1]]);
-
     return {
-      points,
-      residuals,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA}x^${coeffB}`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -293,9 +223,6 @@ const methods = {
       ),
     ]);
 
-    const points = data.map((point) => predict(point[0]));
-    const residuals = points.map((point, index) => [point[0], point[1] - data[index][1]]);
-
     let string = 'y = ';
     for (let i = coefficients.length - 1; i >= 0; i--) {
       if (i > 1) {
@@ -309,11 +236,9 @@ const methods = {
 
     return {
       string,
-      points,
-      residuals,
       predict,
       equation: [...coefficients].reverse(),
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 };
